Add 'City' column and 'Delete' button in UsersPage
A new 'City' column has been added to the data grid in the UsersPage which displays the city information from a user's address. Also, a 'Delete' button is introduced under the 'Action' column, providing an additional functionality to the user interface. Enhancements in styling and indentation for improved code readability were also made.